Remove node -1:-1 as a special case.
Push this value onto the stack like any other node. Fix printing
of "Sum" line to allow us to use "G[-1]" for "\G-1:-1".

This change also means that we need to add elements P[-1] and G[-1]
to the P and G arrays, respectively.

This means I now have a completely uniform way of handling all of
the nodes. There are essentially no special cases.