Move utility routines into a separate file.
The modules Gij, PijGij, and Sum we being created for every adder.
There are often multiple adders which become part of a circuit and
the compiler to compiling these utility routines multiple times and
warning of the duplicate modules.

To clean things up the PijGij.v holds these 3 utility modules in a
separate file.